requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
requests>=2.32.0
pandas>=2.2.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
httpx>=0.27.0
pytest>=8.0.0
